move delta backend from execution prover to custom api
